added new api feature

+++ src/main/java/com/main/weatherman/model/City.java
@@ -18,9 +18,9 @@ public class City {
 
     private String name;
 
-    private float lat;
+    private double lat;
 
-    private float lon;
+    private double lon;
 
     @ManyToOne
     @JoinColumn(name = "country_id")
@@ -38,19 +38,19 @@ public class City {
         this.name = name;
     }
 
-    public float getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(float lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public float getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(float lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
